+++ packages/ui/src/components/data-display/index.ts
@@ -14,3 +14,4 @@ export * from './Table';
 export * from './Text';
 export * from './VariableWeightText';
 export * from './Typewriter';
+export * from './OpenGraphCard';